moved book list to auth screens
