+++ wolox-books/src/app/app.module.ts
@@ -9,12 +9,12 @@ import { SignUpComponent } from './screens/unauth/screens/sign-up/sign-up.compon
 import { LogInComponent } from './screens/unauth/screens/log-in/log-in.component'
 import { AuthComponent } from './screens/auth/auth.component';
 import { LocalStorageService } from './services/local-storage.service';
-import { BookListComponent } from './screens/book-list/book-list.component';
+import { BookListComponent } from './screens/auth/screens/book-list/book-list.component';
 import { UnauthComponent } from './screens/unauth/unauth.component'
 import { AuthGuard } from './auth.guard';
 import { TokenInterceptorService } from './services/token-interceptor.service';
 import { CommonModule } from '@angular/common';
-import { BookCardComponent } from './screens/book-list/components/book-card/book-card.component';
+import { BookCardComponent } from './screens/auth/screens/book-list/components/book-card/book-card.component';
 
 @NgModule({
   declarations: [
