+++ wolox-books/src/app/screens/auth/screens/book-list/components/book-card/book-card.component.html
